+++ gtsfm/scene_optimizer.py
@@ -16,9 +16,9 @@ import numpy as np
 from dask.delayed import Delayed
 
 import gtsfm.evaluation.metrics_report as metrics_report
+import gtsfm.two_view_estimator
 import gtsfm.utils.io as io_utils
 import gtsfm.utils.logger as logger_utils
-import gtsfm.utils.metrics as metrics_utils
 import gtsfm.utils.viz as viz_utils
 from gtsfm.common.image import Image
 from gtsfm.feature_extractor import FeatureExtractor
@@ -150,11 +150,11 @@ class SceneOptimizer:
 
         # persist all front-end metrics and its summary
         auxiliary_graph_list.append(
-            dask.delayed(metrics_utils.persist_frontend_metrics_full)(two_view_reports_dict, image_graph)
+            dask.delayed(io_utils.save_full_frontend_metrics)(two_view_reports_dict, image_graph)
         )
         if gt_pose_graph is not None:
             metrics_graph_list.append(
-                dask.delayed(metrics_utils.aggregate_frontend_metrics)(
+                dask.delayed(two_view_estimator.aggregate_frontend_metrics)(
                     two_view_reports_dict, self._pose_angular_error_thresh
                 )
             )
@@ -276,9 +276,9 @@ def save_metrics_reports(metrics_graph: Delayed) -> List[Delayed]:
         return save_metrics_list
 
     # Save metrics to JSON
-    save_metrics_graph_list.append(dask.delayed(metrics_utils.save_metrics_as_json)(metrics_graph_list, METRICS_PATH))
+    save_metrics_graph_list.append(dask.delayed(io_utils.save_metrics_as_json)(metrics_graph_list, METRICS_PATH))
     save_metrics_graph_list.append(
-        dask.delayed(metrics_utils.save_metrics_as_json)(metrics_graph_list, REACT_METRICS_PATH)
+        dask.delayed(io_utils.save_metrics_as_json)(metrics_graph_list, REACT_METRICS_PATH)
     )
     save_metrics_graph_list.append(
         dask.delayed(metrics_report.generate_metrics_report_html)(
