+++ gtsfm/utils/metrics.py
@@ -7,7 +7,6 @@ from pathlib import Path
 from typing import Dict, List, Optional, Tuple, Union
 
 import numpy as np
-from dask.delayed import Delayed
 from gtsam import Cal3Bundler, EssentialMatrix, Point3, Pose3, Rot3, Unit3
 
 import gtsfm.utils.geometry_comparisons as comp_utils
@@ -17,7 +16,6 @@ import gtsfm.utils.verification as verification_utils
 from gtsfm.common.image import Image
 from gtsfm.common.keypoints import Keypoints
 from gtsfm.evaluation.metrics import GtsfmMetric, GtsfmMetricsGroup
-from gtsfm.two_view_estimator import TwoViewEstimationReport
 
 REPO_ROOT = Path(__file__).resolve().parent.parent.parent
 
@@ -233,121 +231,3 @@ def log_sfm_summary() -> None:
         "Averaging median_trans_dist_err: %.2f", averaging_metrics["translation_averaging_distance"]["median_error"]
     )
     logger.info("Averaging max_trans_dist_err: %.2f", averaging_metrics["translation_averaging_distance"]["max_error"])
-
-
-def persist_frontend_metrics_full(
-    two_view_report_dict: Dict[Tuple[int, int], TwoViewEstimationReport], images: List[Image]
-) -> None:
-    """Persist the front-end metrics for every pair on disk.
-
-    Args:
-        two_view_report_dict: front-end metrics for pairs of images.
-        images: list of all images for this scene, in order of image/frame index.
-    """
-    metrics_list = []
-
-    for (i1, i2), report in two_view_report_dict.items():
-
-        # Note: if GT is unknown, then R_error_deg, U_error_deg, and inlier_ratio_gt_model will be None
-        metrics_list.append(
-            {
-                "i1": i1,
-                "i2": i2,
-                "i1_filename": images[i1].file_name,
-                "i2_filename": images[i2].file_name,
-                "rotation_angular_error": round(report.R_error_deg, PRINT_NUM_SIG_FIGS) if report.R_error_deg else None,
-                "translation_angular_error": round(report.U_error_deg, PRINT_NUM_SIG_FIGS)
-                if report.U_error_deg
-                else None,
-                "num_inliers_gt_model": report.num_inliers_gt_model if report.num_inliers_gt_model else None,
-                "inlier_ratio_gt_model": round(report.inlier_ratio_gt_model, PRINT_NUM_SIG_FIGS)
-                if report.inlier_ratio_gt_model
-                else None,
-                "inlier_ratio_est_model": round(report.inlier_ratio_est_model, PRINT_NUM_SIG_FIGS),
-                "num_inliers_est_model": report.num_inliers_est_model,
-            }
-        )
-
-    io_utils.save_json_file(os.path.join(METRICS_PATH, "frontend_full.json"), metrics_list)
-
-    # Save duplicate copy of 'frontend_full.json' within React Folder.
-    io_utils.save_json_file(os.path.join(REACT_METRICS_PATH, "frontend_full.json"), metrics_list)
-
-
-def aggregate_frontend_metrics(
-    two_view_reports_dict: Dict[Tuple[int, int], TwoViewEstimationReport], angular_err_threshold_deg: float
-) -> None:
-    """Aggregate the front-end metrics to log summary statistics.
-
-    We define "pose error" as the maximum of the angular errors in rotation and translation, per:
-        SuperGlue, CVPR 2020: https://arxiv.org/pdf/1911.11763.pdf
-        Learning to find good correspondences. CVPR 2018:
-        OA-Net, ICCV 2019:
-        NG-RANSAC, ICCV 2019:
-
-    Args:
-        two_view_report_dict: report containing front-end metrics for each image pair.
-        angular_err_threshold_deg: threshold for classifying angular error metrics as success.
-    """
-    num_image_pairs = len(two_view_reports_dict.keys())
-
-    # all rotational errors in degrees
-    rot3_angular_errors = [report.R_error_deg for report in two_view_reports_dict.values()]
-    trans_angular_errors = [report.U_error_deg for report in two_view_reports_dict.values()]
-    rot3_angular_errors = np.array(rot3_angular_errors, dtype=float)
-    trans_angular_errors = np.array(trans_angular_errors, dtype=float)
-    # count number of rot3 errors which are not None. Should be same in rot3/unit3
-    num_valid_image_pairs = np.count_nonzero(~np.isnan(rot3_angular_errors))
-
-    # compute pose errors by picking the max error from rot3 and unit3 errors
-    pose_errors = np.maximum(rot3_angular_errors, trans_angular_errors)
-
-    # check errors against the threshold
-    success_count_rot3 = np.sum(rot3_angular_errors < angular_err_threshold_deg)
-    success_count_unit3 = np.sum(trans_angular_errors < angular_err_threshold_deg)
-    success_count_pose = np.sum(pose_errors < angular_err_threshold_deg)
-
-    # count image pair entries where inlier ratio w.r.t. GT model == 1.
-    all_correct = np.count_nonzero([report.inlier_ratio_gt_model == 1.0 for report in two_view_reports_dict.values()])
-
-    logger.debug(
-        "[Two view optimizer] [Summary] Rotation success: %d/%d/%d",
-        success_count_rot3,
-        num_valid_image_pairs,
-        num_image_pairs,
-    )
-
-    logger.debug(
-        "[Two view optimizer] [Summary] Translation success: %d/%d/%d",
-        success_count_unit3,
-        num_valid_image_pairs,
-        num_image_pairs,
-    )
-
-    logger.debug(
-        "[Two view optimizer] [Summary] Pose success: %d/%d/%d",
-        success_count_pose,
-        num_valid_image_pairs,
-        num_image_pairs,
-    )
-
-    logger.debug(
-        "[Two view optimizer] [Summary] # Image pairs with 100%% inlier ratio:: %d/%d", all_correct, num_image_pairs
-    )
-
-    #TODO(akshay-krishnan): Move angular_err_threshold_deg and num_total_image_pairs to metadata.
-    frontend_metrics = GtsfmMetricsGroup("frontend_summary", [
-        GtsfmMetric("angular_err_threshold_deg", angular_err_threshold_deg), 
-        GtsfmMetric("num_total_image_pairs", int(num_image_pairs)),
-        GtsfmMetric("num_valid_image_pairs", int(num_valid_image_pairs)), 
-        GtsfmMetric("rotation_success_count", int(success_count_rot3)), 
-        GtsfmMetric("translation_success_count", int(success_count_unit3)), 
-        GtsfmMetric("pose_success_count", int(success_count_pose)), 
-        GtsfmMetric("num_all_inlier_correspondences_wrt_gt_model", int(all_correct))
-    ])
-    return frontend_metrics
-
-
-def save_metrics_as_json(metrics_groups: Delayed, output_dir: str) -> None:
-    for metrics_group in metrics_groups:
-        metrics_group.save_to_json(os.path.join(output_dir, metrics_group.name + ".json"))
